# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LeadGen API",
    "description": "API for generating business leads from Google Maps with email enrichment",
    "version": "1.0.0"
  },
  "paths": {
    "/v1/maps/search": {
      "get": {
        "tags": [
          "maps"
        ],
        "summary": "Search Businesses",
        "operationId": "search_businesses_v1_maps_search_get",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Business type or keyword (e.g. 'dentistas')",
              "title": "Query"
            },
            "description": "Business type or keyword (e.g. 'dentistas')"
          },
          {
            "name": "location",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Location (e.g. 'Madrid, Spain')",
              "default": "",
              "title": "Location"
            },
            "description": "Location (e.g. 'Madrid, Spain')"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Max results",
              "default": 20,
              "title": "Limit"
            },
            "description": "Max results"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/maps/place": {
      "get": {
        "tags": [
          "maps"
        ],
        "summary": "Get Place",
        "operationId": "get_place_v1_maps_place_get",
        "parameters": [
          {
            "name": "place_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Google Maps Place ID",
              "title": "Place Id"
            },
            "description": "Google Maps Place ID"
          },
          {
            "name": "enrich",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Enrich with email and social data",
              "default": false,
              "title": "Enrich"
            },
            "description": "Enrich with email and social data"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/health": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Health Check",
        "operationId": "health_check_v1_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/": {
      "get": {
        "summary": "Root",
        "operationId": "root__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BusinessResult": {
        "properties": {
          "place_id": {
            "type": "string",
            "title": "Place Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "website": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Website"
          },
          "rating": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating"
          },
          "reviews_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviews Count"
          },
          "coordinates": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Coordinates"
              },
              {
                "type": "null"
              }
            ]
          },
          "hours": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Hours"
              },
              {
                "type": "null"
              }
            ]
          },
          "is_claimed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Claimed"
          },
          "enrichment": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Enrichment"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "place_id",
          "name"
        ],
        "title": "BusinessResult"
      },
      "Coordinates": {
        "properties": {
          "lat": {
            "type": "number",
            "title": "Lat"
          },
          "lng": {
            "type": "number",
            "title": "Lng"
          }
        },
        "type": "object",
        "required": [
          "lat",
          "lng"
        ],
        "title": "Coordinates"
      },
      "EmailInfo": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "type": {
            "type": "string",
            "title": "Type"
          }
        },
        "type": "object",
        "required": [
          "email",
          "status",
          "type"
        ],
        "title": "EmailInfo"
      },
      "Enrichment": {
        "properties": {
          "emails": {
            "items": {
              "$ref": "#/components/schemas/EmailInfo"
            },
            "type": "array",
            "title": "Emails",
            "default": []
          },
          "social": {
            "$ref": "#/components/schemas/SocialLinks",
            "default": {}
          }
        },
        "type": "object",
        "title": "Enrichment"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Hours": {
        "properties": {
          "monday": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Monday"
          },
          "tuesday": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tuesday"
          },
          "wednesday": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wednesday"
          },
          "thursday": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thursday"
          },
          "friday": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Friday"
          },
          "saturday": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Saturday"
          },
          "sunday": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sunday"
          }
        },
        "type": "object",
        "title": "Hours"
      },
      "Pagination": {
        "properties": {
          "current_page": {
            "type": "integer",
            "title": "Current Page"
          },
          "total_results": {
            "type": "integer",
            "title": "Total Results"
          },
          "has_next": {
            "type": "boolean",
            "title": "Has Next"
          }
        },
        "type": "object",
        "required": [
          "current_page",
          "total_results",
          "has_next"
        ],
        "title": "Pagination"
      },
      "SearchResponse": {
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/BusinessResult"
            },
            "type": "array",
            "title": "Results"
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        },
        "type": "object",
        "required": [
          "results",
          "pagination"
        ],
        "title": "SearchResponse"
      },
      "SocialLinks": {
        "properties": {
          "linkedin": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Linkedin"
          },
          "instagram": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instagram"
          },
          "facebook": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Facebook"
          },
          "twitter": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Twitter"
          },
          "youtube": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Youtube"
          }
        },
        "type": "object",
        "title": "SocialLinks"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}